requirements.txt
torch>=2.2
torchvision>=0.17
diffusers>=0.30
datasets>=2.20
Pillow>=10.0
tqdm>=4.66
accelerate>=0.33
